requirements.txt
# Core automation dependencies
arxiv==2.1.0
pyyaml==6.0.1

# Google Drive integration
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# Data processing (useful for analysis)
pandas==2.1.4
numpy==1.26.2

# Optional but recommended
requests==2.31.0
tqdm==4.66.1
